// src/theme/colors.type.ts
export type ColorType = {
  background: string;
  overlay: string;
  surface: string;
  onSurface: string;
  onBackground: string;

  primary: string;
  onPrimary: string;

  secondary: string;
  onSecondary: string;

  textLight: string;
  textDark: string;

  border: string;

  error: string;
  success: string;
  warning: string;
};